--- lib/core/terraform/config/base.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "dsl"
-
-module Terraform
-  module Config
-    class Base
-      include Dsl
-
-      def to_tf
-        raise NotImplementedError
-      end
-    end
-  end
-end
