+++ lib/core/terraform_config/base.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require_relative "dsl"
+
+module TerraformConfig
+  class Base
+    include Dsl
+
+    def to_tf
+      raise NotImplementedError
+    end
+  end
+end
